AI-Creation-Roadmap-Python-StepByStep/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
sqlalchemy==2.0.35
psycopg[binary,pool]==3.2.3
pgvector==0.3.4
numpy==2.1.1
scikit-learn==1.5.2
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.25
rank-bm25==0.2.2
